+++ src/Entities/Enums/DataResidency.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace GlobalPayments\Api\Entities\Enums;
+
+use GlobalPayments\Api\Entities\Enum;
+
+class DataResidency extends Enum
+{
+    const EU = "EU";
+    const NONE = "NONE";
+}